Move SerializationResourceManager to own file
